# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AttendanceResponse:
    properties:
      date:
        type: string
      discipline:
        type: string
      lesson_id:
        type: integer
      recorded_at:
        type: string
      start_time:
        type: string
      status:
        type: string
      teacher:
        type: string
    type: object
  dto.AttendanceStatsResponse:
    properties:
      absent:
        type: integer
      attendance_rate:
        type: number
      attended:
        type: integer
      late:
        type: integer
      period_end:
        type: string
      period_start:
        type: string
      student_id:
        type: integer
      total_lessons:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
  dto.GradeResponse:
    properties:
      created_at:
        type: string
      discipline:
        type: string
      discipline_id:
        type: integer
      grade_type:
        type: string
      grade_value:
        type: number
      id:
        type: integer
      semester:
        type: string
    type: object
  dto.GroupAttendanceSummary:
    properties:
      attendance_rate:
        type: number
      attended:
        type: integer
      student_id:
        type: integer
      student_name:
        type: string
      total_lessons:
        type: integer
    type: object
  dto.GroupDetailsResponse:
    properties:
      admission_year:
        type: integer
      faculty:
        type: string
      id:
        type: integer
      name:
        type: string
      specialization:
        type: string
      student_count:
        type: integer
      students:
        items:
          $ref: '#/definitions/dto.GroupStudentResponse'
        type: array
    type: object
  dto.GroupResponse:
    properties:
      admission_year:
        type: integer
      faculty:
        type: string
      id:
        type: integer
      name:
        type: string
      specialization:
        type: string
      student_count:
        type: integer
    type: object
  dto.GroupStudentResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      student_id_number:
        type: string
    type: object
  dto.LessonAttendanceResponse:
    properties:
      attendance:
        items:
          $ref: '#/definitions/dto.StudentAttendanceDetail'
        type: array
      date:
        type: string
      discipline:
        type: string
      lesson_id:
        type: integer
    type: object
...
  dto.MarkAttendanceRequest:
    properties:
      lesson_id:
        type: integer
      status:
        enum:
        - present
        - absent
        - late
        type: string
      student_id:
        type: integer
    required:
    - lesson_id
    - status
    - student_id
    type: object
  dto.StudentAcademicResponse:
    properties:
      average_grade:
        type: number
      completed_credits:
        type: integer
      current_semester:
        type: string
      grades:
        items:
          $ref: '#/definitions/dto.GradeResponse'
        type: array
      student_id:
        type: integer
      total_disciplines:
        type: integer
    type: object
  dto.StudentAttendanceDetail:
    properties:
      recorded_at:
        type: string
      status:
        type: string
      student_id:
        type: integer
      student_id_number:
        type: string
      student_name:
        type: string
    type: object
  dto.StudentProfileResponse:
    properties:
      admission_year:
        type: integer
      email:
        type: string
      faculty_id:
        type: integer
      faculty_name:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      id:
        type: integer
      name:
        type: string
      specialization:
        type: string
      student_id_number:
        type: string
      user_id:
        type: integer
    type: object
  dto.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для системы расписания студентов
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Student Schedule API
  version: "1.0"
paths:
  /attendance/groups/{id}/summary:
    get:
      consumes:
      - application/json
      description: Возвращает сводку посещаемости всех студентов группы за период
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: 'Начальная дата (формат: 2006-01-02)'
        example: '"2024-01-01"'
        in: query
        name: start_date
        type: string
      - description: 'Конечная дата (формат: 2006-01-02)'
        example: '"2024-01-31"'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GroupAttendanceSummary'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить сводку посещаемости по группе
      tags:
      - attendance
  /attendance/lessons/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает посещаемость конкретного занятия
      parameters:
      - description: ID занятия
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonAttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить посещаемость занятия
      tags:
      - attendance
  /attendance/mark:
    post:
      consumes:
      - application/json
      description: Отмечает посещение студента на занятии
      parameters:
      - description: Данные посещения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MarkAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Отметить посещение
      tags:
      - attendance
  /attendance/mark-batch:
    post:
      consumes:
      - application/json
      description: Отмечает посещение для нескольких студентов на занятии
      parameters:
      - description: Данные посещений
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.MarkAttendanceRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Отметить посещение для нескольких студентов
      tags:
      - attendance
  /attendance/students/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает посещаемость студента за указанный период
      parameters:
      - description: ID студента
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: 'Начальная дата (формат: 2006-01-02)'
        example: '"2024-01-01"'
        in: query
        name: start_date
        type: string
      - description: 'Конечная дата (формат: 2006-01-02)'
        example: '"2024-01-31"'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AttendanceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить посещаемость студента
      tags:
      - attendance
  /attendance/students/{id}/stats:
    get:
      consumes:
      - application/json
      description: Возвращает статистику посещаемости студента за период
      parameters:
      - description: ID студента
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - default: week
        description: Период (week, month, semester)
        enum:
        - week
        - month
        - semester
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttendanceStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить статистику посещаемости
      tags:
      - attendance
  /groups:
    get:
      consumes:
      - application/json
      description: Возвращает список всех групп
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GroupResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить все группы
      tags:
      - groups
  /groups/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о группе включая студентов
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GroupDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить детали группы
      tags:
      - groups
  /groups/faculty/{faculty_id}:
    get:
      consumes:
      - application/json
      description: Возвращает список групп по ID факультета
      parameters:
      - description: ID факультета
        example: 1
        in: path
        name: faculty_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GroupResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить группы по факультету
      tags:
      - groups
...
  /students/{id}/academic:
    get:
      consumes:
      - application/json
      description: Возвращает академическую информацию о студенте (оценки, посещаемость
        и т.д.)
      parameters:
      - description: ID студента
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudentAcademicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить академическую информацию
      tags:
      - students
  /students/{id}/profile:
    get:
      consumes:
      - application/json
      description: Возвращает профиль студента с основной информацией
      parameters:
      - description: ID студента
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudentProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить профиль студента
      tags:
      - students
schemes:
- http
swagger: "2.0"